# small optimization