--- a/lib/cli-entry.js
+++ b/lib/cli-entry.js
@@ -4,6 +4,7 @@
 module.exports = async (process, validateEngines) => {
   // set it here so that regardless of what happens later, we don't
   // leak any private CLI configs to other programs
+  process.title = 'npm'
 
   // if npm is called as "npmg" or "npm_g", then run in global mode.
   if (process.argv[1][process.argv[1].length - 1] === 'g') {
@@ -56,7 +57,6 @@ module.exports = async (process, validateEngines) => {
       process.exitCode = 1
       return exitHandler()
     }
-    setTimeout(() => process.title = npm.title)
     await npm.exec(cmd)
     return exitHandler()
   } catch (err) {