--- a/lib/npm.js
+++ b/lib/npm.js
@@ -12,7 +12,6 @@ const LogFile = require('./utils/log-file.js')
 const Timers = require('./utils/timers.js')
 const Display = require('./utils/display.js')
 const log = require('./utils/log-shim')
-const replaceInfo = require('./utils/replace-info.js')
 const updateNotifier = require('./utils/update-notifier.js')
 const pkg = require('../package.json')
 const { deref } = require('./utils/cmd-list.js')
@@ -177,7 +176,6 @@ class Npm {
   }
 
   set title (t) {
-    process.title = t
     this.#title = t
   }
 
@@ -227,6 +225,7 @@ class Npm {
     // note: this MUST be shorter than the actual argv length, because it
     // uses the same memory, so node will truncate it if it's too long.
     this.time('npm:load:setTitle', () => {
+      const replaceInfo = require('./utils/replace-info.js')
       const { parsedArgv: { cooked, remain } } = this.config
       this.argv = remain
       // Secrets are mostly in configs, so title is set using only the positional args