--- a/lib/npm.js
+++ b/lib/npm.js
@@ -176,6 +176,7 @@ class Npm {
   }
 
   set title (t) {
+    process.title = t
     this.#title = t
   }