--- a/lib/utils/replace-info.js
+++ b/lib/utils/replace-info.js
@@ -1,4 +1,4 @@
-const { cleanUrl } = require('npm-registry-fetch')
+const cleanUrl = require('npm-registry-fetch/lib/clean-url')
 const isString = (v) => typeof v === 'string'
 
 // split on \s|= similar to how nopt parses options