@sirenkovladd 2024-01-24
```sh
❯ hyperfine -w 5 -r 20 'npm run echo' 'node bin/npm-cli.js run echo'
Benchmark 1: npm run echo
  Time (mean ± σ):     253.7 ms ±   4.0 ms    [User: 252.9 ms, System: 34.0 ms]
  Range (min … max):   250.2 ms … 268.7 ms    20 runs
 
Benchmark 2: node bin/npm-cli.js run echo
  Time (mean ± σ):     208.4 ms ±   5.0 ms    [User: 184.4 ms, System: 24.1 ms]
  Range (min … max):   205.9 ms … 229.5 ms    20 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  node bin/npm-cli.js run echo ran
    1.22 ± 0.04 times faster than npm run echo
```