@wraithgar 2024-04-09
This has been addressed by https://github.com/npm/cli/pull/7334